Rename Database and DatabaseConfig.py
